numba version of RCS computation
Created a numba version of the rcs computation. This is now used by default, but the non-numba code is still available (set numba=False in init, or use rcs_classic). It runs faster for big scenes. For small scenes the improvement is marginal. Further optimization of the code should be possible.